requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
google-generativeai
chromadb
sentence-transformers
pypdf
pydantic
twilio
httpx
python-multipart
openai-whisper
gTTS
streamlit
sounddevice
scipy
python-dotenv
docx2txt
langchain-google-genai
